requirements.txt
fastapi
uvicorn[standard]
aiomqtt
python-multipart
python-dotenv
